Crate fasten_combine

Collapse identical reads into single reads, recalculating quality values. If paired end, then each set of reads must be identical to be collapsed. Warning: due to multiple reads collapsing into one, read identifiers will be reconstituted.

Examples

cat testdata/four_reads | fasten_combine > combined.fastq

Usage

Usage: fasten_combine [-h] [-n INT] [-p] [-v] [--max-qual-char CHAR] [--min-qual-char CHAR]

Options:
    -h, --help          Print this help menu.
    -n, --numcpus INT   Number of CPUs (default: 1)
    -p, --paired-end    The input reads are interleaved paired-end
    -v, --verbose       Print more status messages
        --max-qual-char CHAR
                        Maximum quality character (default: I)
        --min-qual-char CHAR
                        Minimum quality character (default: !)

    NOTE: range of quality scores is !"#$%&'()*+,-./0123456789:;<=>?@ABCDEFGHI

Constants

READ_SEPARATOR 🔒

Glues together paired end reads internally and is a character not expected in any read

TEN 🔒

need this constant because the compiler had a problem with the syntax 10.0.pow()

Functions

combine_error_vectors 🔒

Combines vectors of error probabilities such that the rate of error is probability of error from vector one times the probability of error from vector two.

main 🔒